feat: 연차 영속 — 내 연차 조회·수정·사용내역 (#89) (#108)
* feat: 연차 영속 — 내 연차 조회·수정·사용내역 (#89)

와이어프레임 "내 연차" 화면과 홈 남은연차 배지가 백엔드 공백이었다.
GET /home 의 remainingLeaveDays 는 클라이언트가 넘긴 값을 그대로 되돌려줄
뿐이었고(HomeService.home(Integer) → new HomeResult(그 값, cards)), 연차를
저장하는 계층이 아예 없었다.

남은 연차를 저장하지 않는다. 총 연차(leave_balance)와 증감 내역(leave_usage)만
두고 남은 값은 LeaveSummary 가 파생한다 — 남은 값을 따로 저장하면 내역과
어긋난다(차감은 됐는데 갱신이 실패하는 식). 내역이 정본이면 언제든 다시
계산해 맞출 수 있다.

내역은 '증감' 이다. 코스를 취소하면 음수 내역을 하나 더 쌓아 되돌린다 — 행을
지우면 언제 무엇이 취소됐는지가 사라진다. 대신 0 은 막는다(아무것도 바꾸지
않는 기록은 소음이다).

남은 연차는 음수가 될 수 있다. 초과 사용을 서버가 막지 않기 때문이다(결정
#38) — 0 으로 깎으면 얼마나 초과했는지를 화면이 알 수 없다.

엔드포인트 (소유 키는 코스와 같은 X-Guest-Id)
- GET   /api/v1/leaves/me        현황 + 내역. 설정한 적 없으면 총 0(404 아님)
- PATCH /api/v1/leaves/me        총 연차 수정(0.5 단위, 0~365)
- POST  /api/v1/leaves/me/usages 내역 추가 → 201

홈은 저장값을 읽는다. remainingLeave 쿼리 파라미터를 없애고 X-Guest-Id
헤더(선택)로 바꿨다. 헤더가 없으면 남은 연차는 null 이고 나머지 홈은 그대로
뜬다. 반차가 0.5 라 타입도 Integer 에서 Double 로 바꿨다.

에러코드 LEAVE-009·010 추가(append-only).

* test: 연차 영속 단위·통합 테스트 + 홈 회귀 대응 (#89)

단위 — 0.5 단위 규칙과 남은 연차 파생. 반차가 0.5 라 0.3 같은 값이 통과하면
이후 합·차감이 조용히 이상해진다. 음수 증감은 허용하고(취소) 0 은 막는다.

통합 — 미설정 소유자는 총 0 으로 답한다(404 아님, 클라이언트가 '처음 쓰는
사람' 을 예외로 다루지 않게), 수정 시 새 행이 생기지 않음, 취소가 음수 내역으로
쌓여 되돌아옴, 반차 0.5, 초과 사용 시 음수 허용, 소유 키 격리, 계약 위반
400 3종.

소유자를 테스트마다 다르게 쓴다 — 이 클래스는 DB 를 롤백하지 않아(공유 컨텍스트)
같은 키면 앞 테스트 잔여 상태가 다음 시나리오로 샌다.

홈 회귀 — remainingLeave 쿼리 파라미터가 X-Guest-Id 헤더로 바뀌었다. 홈
테스트는 이제 PATCH 로 먼저 저장한 뒤 GET 으로 확인해, '저장 → 홈이 읽음'
배선까지 함께 검증한다. 벤치마크 E2E 는 연차와 무관한 측정이라 저장하지 않은
소유 키를 넘긴다.

* fix: 빈 소유 키가 500 을 내던 것 · 도메인 계약 예외 · 동시 생성 경합 (#89)

CodeRabbit 리뷰 대응 3건.

빈 소유 키가 500 이었다. X-Guest-Id: "  " 는 헤더가 '있긴 있어서'
@RequestHeader 를 통과하고, 도메인 불변식에서 터져 COMMON-500 이 나갔다.
실호출로 확인했다 — GET 은 200(조회만 해서 도메인을 안 만든다), PATCH·POST 는
500. 클라이언트 계약 위반이 서버 버그로 보고되던 자리다.

LEAVE-011 을 추가하고 서비스 진입에서 검증한다. 조회에도 건다 — 조회만
실패하지 않는다고 통과시키면 같은 요청이 메서드에 따라 200 과 400 으로 갈린다.

도메인도 계약 예외로 바꿨다. LeaveBalance.requireOwner 와
LeaveUsage.requireDays 가 IllegalArgumentException 을 던졌는데, 지금은 DTO 가
먼저 걸러 안 보이지만 #91 이 forCourse() 를 서비스에서 직접 부르면 DTO 를
거치지 않고 들어온다. 그때 400 대신 500 이 나간다.

동시 생성 경합을 한 곳으로 모았다. changeTotalDays 와 addUsage 가 각자
"조회 후 없으면 생성" 을 하고 있어, 같은 소유자로 요청이 겹치면 둘 다 "없음" 을
보고 만들려다 uk_leave_balance_guest 에 걸려 DataIntegrityViolationException
으로 튕겼다. 먼저 넣은 쪽이 이겼다는 뜻일 뿐이라 잡고 재조회해 다시 적용한다.
upsertBalance 하나가 두 경로를 함께 처리한다.

* fix: 동시 생성이 rollback-only 로 번지던 것 — 트랜잭션 경계 분리 (#89)

CodeRabbit 재리뷰 대응(Major). 지적이 맞았고, 테스트로 재현됐다.

앞 커밋에서 changeTotalDays 의 @Transactional 만 뗐는데 클래스 레벨에
@Transactional(readOnly = true) 가 남아 있었다. 그래서 여전히 바깥 트랜잭션
안이었고, 제약 위반이 그 트랜잭션을 rollback-only 로 찍어 커밋에서
UnexpectedRollbackException 이 났다 — 예외를 잡아도 소용없다.

동시 PATCH 회귀 테스트가 [200, 500] 로 실제로 잡아냈다.

- 클래스 레벨 트랜잭션을 없애고 읽기 메서드에만 개별로 건다
- create() 는 REQUIRES_NEW — 호출자가 어쩌다 트랜잭션 안이어도 실패가
  그쪽으로 번지지 않는다
- addUsage 는 잔액 행을 아예 만들지 않는다. summaryOf 가 없을 때 총 0 으로
  답하므로 결과가 같은데, 미리 만들면 경합 자리만 하나 더 생긴다

동시 요청 테스트는 경합 구간이 좁아 매번 재현되진 않지만, 이번엔 실제로
잡았고 3회 반복 통과를 확인했다.

+++ src/main/java/com/offway/core/trip/service/dto/HomeResult.java
@@ -12,10 +12,10 @@ import java.util.List;
  *
  * <p>홈 카드는 대표 이미지·볼거리 categories 와 대표 혜택 하나를 붙인다(뱃지 강조). 무드 필터는 추천(S3)에서만 건다.
  *
- * @param remainingLeaveDays 남은 연차 (게스트 — 클라이언트가 넘긴 값, 없으면 null)
+ * @param remainingLeaveDays 남은 연차 (저장값. 설정한 적 없으면 null — 0 과 구분한다). 반차가 0.5 라 실수다
  * @param regions 추천 지역 카드 (랭킹 top-N)
  */
-public record HomeResult(Integer remainingLeaveDays, List<RegionCard> regions) {
+public record HomeResult(Double remainingLeaveDays, List<RegionCard> regions) {
 
     /**
      * @param regionId 지역 ID
